+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0-alpha] - 2017-6-5
+## [1.1.1-alpha] - 2017-06-19
+## Changed
+- Fixed lists positions so a user cannot change it to a larger number than the total number of lists they have
+- Fixed lowercase function on the wrong user input on editing lists
+
+## [1.1.0-alpha] - 2017-06-05
 ### Added
 - Uploaded user avatars
 - Image helper class
@@ -15,5 +20,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Lists: force category inputs in lowercase
 - Replaced Markdown library with CommonMark
 
-## [1.0.0-alpha] - 2017-5-30
+## [1.0.0-alpha] - 2017-05-30
 First development release.